+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/SideCar.java
@@ -32,4 +32,11 @@ public class SideCar {
     @Setter
     @Getter
     private String internalErrorCode = "";
+
+    @Setter
+    @Getter
+    private long internalRequestLatencyNanos;
+    @Setter
+    @Getter
+    private long internalResponseLatencyNanos;
 }