+++ docs/en/changes/changes.md
@@ -46,6 +46,7 @@
 * Enhance Deprecated MetricQuery(v1) getValues querying to asynchronous concurrency query
 * Fix the pod match error when the service has multiple selector in kubernetes environment.
 * VM monitoring adapts the 0.50.0 of the `opentelemetry-collector`.
+* Add Envoy internal cost metrics.
 
 #### UI